Jit compilative functions "call" & "ret" implemented
Command "call arg" performs jump to the "arg" label and
pushes adress of itself to the stack. Command "ret" pops
number from stack (it have to be adress of last call, it means
that calling function have to allign stack after itself) and
performs jump to the adress which corresponding popped number.