Now you don't have to make equal pointers for equal strings. And language declaration's size now is not fixed.
Dirty place in JitCompiler.cpp was cleaned. That place
was bound with compare system in std::map. I need to compare
two strings (contained in const char*) but std::map compares
pointers. Because of it i have to make pointers to equal
strings equal. Now i fixed it. Fixed size masif of CommandDeclaration
objects in class LanguageDeclaration was changed to std::map.