Implemented asm labels and Jmp to label
When label binds to some place in code it is
appropriated to some value which is number
of first byte of next instruction after label
in JIT-compilating function. During emit
jmp instruction adress which have to be jumped
assigns 0x0. During make result function these
places changes to absolute adresses. It is
necessary because of variable adresses of elements
of std::vectors.